LibUnicode: Generate separate tables for Unicode properties
Similar to General Categories, this generates separate tables for the
Property list.

+++ Userland/Libraries/LibUnicode/Forward.h
@@ -13,7 +13,7 @@ namespace Unicode {
 enum class Condition : u8;
 enum class GeneralCategory : u8;
 enum class Locale : u8;
-enum class Property : u64;
+enum class Property : u8;
 enum class Script : u8;
 enum class WordBreakProperty : u8;
